# Feature Issue 102 -- Liststate entry is deleted from the connector state without checking if the list is deleted in state file
# Problem
Webstate.endRecrawl() did not do HTTP 404 check for a list before marking it as deleted
in the webstate for metadata-and-URL feed mode. This even applies to the way web states are marked as invalid

# High level list of changes
1. Updated GlobalState.endReCrawl() to remove the check for content feed that ensures HTTP 404 check is applied even for metadata-and-URL feed.

Any other exceptions or errors will be considered temporary and not be used to
mark the web state as invalid and be deleted.
Earlier this check was only for content feed mode.

# Specific flow/logic that the reviewer needs to focus on
Removed the if() block for checking the content feed mode before the call to
checkConnectivity() and the check that the response code is not HTTP 404