# Feature or Bug fix reference Issue 85: The discovery of site collection should be initiated in the first crawl cycle even though no content is crawled
# Problem description
The discovery of site collections is not initiated as it is set to discover 
after one complete traversal cycle. This causes a problem if the first traversal cycle returns zero docs


# High level list of changes
Updated the updateGlobalState(GlobalState globalState) to use the doCrawl & isGSupdated flags and not bFullRecrawl flag for initiating the discovery of new sites. 
Added a new function traverseSites() that crawls a given list of sites

#  Specific flow/logic that the reviewer needs to focus on
1. if() block : if (doCrawl && null != spType) 

2. Check if the logic handles nextLastWebID and nextLastListID correctly

# Testing: How has the change being tested
Not yet tested. Waiting for initial code review