# Feature or Bug Fix Issue 90 --The time at which connector completes a crawl cycle should be logged in state file
# Problem
Bug: No timestamp was being logged for web states

The logging of timestamp for webstate was overridden in revision 384. Due to this it was not being logged.

# High level list of changes
For webstate the timestamp is related to when was the last liststate queried for
any changes OR when a discovery of child sites was initiated

# Specific flow/logic that the reviewer needs to focus on
source/java/com/google/enterprise/connector/sharepoint/client/SharepointClient.j
ava
1. traversesites()
   Updated to set the timestamp for webstate here
source/java/com/google/enterprise/connector/sharepoint/state/WebState.java


# Testing: How has the change being tested
Manually by running the connector and checking the value of date and time in the
SharePoint_state.xml file