# Feature or Bug Fix  Issue 90  --The time at which connector completes a crawl cycle should be logged in state file
# Problem
Bug: No timestamp was being logged for web states

The logging of timestamp for webstate was overridden in revision 384. Due to
this it was not being logged.

# High level list of changes
Incorporated code review comments from revision 448

# Specific flow/logic that the reviewer needs to focus on
source/java/com/google/enterprise/connector/sharepoint/client/SharepointClient.j
ava

Moved the ws.setLastCrawledDateTime() call in updateWebStateFromSite


# Testing: How has the change being tested
Manually by running the connector and checking the value of date and time in the
SharePoint_state.xml file