# Feature or Bug Fix >> Incorporated review comments on revision 593 >> During the incremental crawl, connector always picks the latest recieved change token for the next web service call. There can be mismatch between the latest change token and the one recieved when first WS call was made with the current change token. This will mainly occur if  the LastDoc and batchint value varies for the same given change token. Connector, in that case, should pick the first recieved change token for the subsequent WS calls as this token is the one which defines the change log window for current traversal.
Though, this scenario is never going to occur in the connector's lifecycle simply becasue the crawl queue always gets empty before the next WS 
call begins and hence there will be no case when a valid LastDoc value will be recieved along with a change token as input parameters for the WS call.

Still, to make the methods implemntation ideal for such scenarios, it should be ensured that the correct input parameters (LastDoc + Change Token) is picked while making WS calls.

# High level list of changes
* Split the change token updating logic in two steps: 
Step1: The next suitable value of change token is saved when WS call returns. This is done only for the first time when the WS call is made corresponding to a new Change Token.
Step2: Commit the next suitable change token value as current change token when all the documents are fed and also there are no more documents to be discovered in the current change log wndow.

# Specific flow/logic that the reviewer needs to focus on
/source/java/com/google/enterprise/connector/sharepoint/wsclient/ListsWS.java
/source/java/com/google/enterprise/connector/sharepoint/state/ListState.java
/source/java/com/google/enterprise/connector/sharepoint/spiimpl/SPDocumentList.java
        - Chnaged the Change Token updation logic
        - A complete iterator for crawl queue is returned for scanning instead of an iterator for sub-list. This is safe becasue at the time of checkpointing all the documents already fed to gsa are removed.
        - getLastDocForWS checks the crawlqueue first before returning the LastDocAsWS which is suitable for making WS calls.

/source/java/com/google/enterprise/connector/sharepoint/spiimpl/SPDocument.java
	- Incoporated review comments on revision 593

/SharePoint Connector Trunk/source/java/com/google/enterprise/connector/sharepoint/client/SharepointClient.java
/SharePoint Connector Trunk/source/java/com/google/enterprise/connector/sharepoint/client/SPConstants.java
/SharePoint Connector Trunk/source/java/com/google/enterprise/connector/sharepoint/state/GlobalState.java
/SharePoint Connector Trunk/source/java/com/google/enterprise/connector/sharepoint/state/WebState.java
        - Others, dependent changes.

# Testing: How has the change being tested
- Using JUnit test cases
- By running the connector end to end