Feature Description ------------------- This check-in contains the fixes for following two bugs reported during RC1 testing of SharePoitn Connector 2.4.6: 1. Zone specific security policies are not considered while constructing ACL 2. Site collection Administrators are sent as users in the ACL. There should be a common SPGroup that should be used for such users
High level list of changes
--------------------------
The fixes for both the bugs are in web service implementation only.
GssAclUtility.FetchSecurityPolicyForAcl() has been updated to consider the zone specific policies while constructing the ACL.
	
GssAclUtility.FetchSiteAdminsForAcl() has been updated so that it will add a single principal (which is a hypothetical SP Group) to the ACL and all the administrator users are set as members of this principal group.
GssAclMonitor.ResolvePrincipal() has been updated to understand the newly added custom SP group for site collection administrator.

The default constructors of some classes have been made private so that the caller always provides values of mandatory fields through the appropriate constructor invocation.  
This is important because when the SOAP response is constructed, no object reference should be null.

LogMessage type is changed to string from StringBuilder. Though, StringBuilder makes perfect sense to have, it does not seems to be fully serializable; the actual string value gets dropped in the response.

The logic of updating change token in the GetAclChangesSinceToken() is improved for better understanding and to make it more safe and reliable.

Specific flow/logic that the reviewer needs to focus on
-------------------------------------------------------
/SharePoint Connector Trunk/source/javatests/com/google/enterprise/connector/sharepoint/wsclient/GssAclTest.java
	-- Contains all the changes

/SharePoint Connector Trunk/source_net/Acl/GssAcl.asmx
	-- Updated JUnit

Rest all files are dependent/generated files for the web service and can be ignored during the review.
	
Testing: How has the change being tested
----------------------------------------
1. Unit testing with Junit test cases