Incorporated comments from r861 and r862
Configuration of filter patterns have been moved to Axis configuration file. Exposing this through ConnectorInstance.xml was misleading since that would give a wrong impression to the user that this configuration is connector instance specific which is not the case.

Filtering invalid references is hard coded as we know it always fails. In addition, more patterns can be provided through Axis configuration.