Fix Details =======================
1) GSASearchArea.ascx 

• Added the following  scopes to the scopes dropdown namely :-

  o Farm
  o Current site
  o Current site & all subsites
  o Current list
  o Current folder
  o Current folder & all sub folders

• The code for retrieving the scope url is now embedded into the file GSASearchArea.ascx. This value is 

then added as the ‘value’ field for the each of the options in the scope dropdown.  

• Added ‘Databinding expression’ (i.e. code within <% and %>)  in the HTML source code for displaying 

the options as per the current context user is browsing. For Eg :- If a user is browsing through a site, 

then the scope dropdown will display only ‘site search’ relative options, and no options would be 

displayed relative to a ‘folder search’. 

2) GSASearchresults.aspx

• Included ‘inurl’ parameter into the search request sent to the GSA

• Modified the ‘SetPageTitle’ javascript to display the selected scope in scope dropdown, if the option 

selected is Current List/ Current Folder/ Current Folder and all subfolders, as these options are not 

visible while browsing a site.  The change is made to temporarily display the selected option on 

GSASearchresults.aspx page, to carry forward the scope.