# Feature or Bug Fix
Contains fix for the issue found when Viewstate is enabled on SharePoint server 2010 at page level.

# High level list of changes

Modified code for javascript function which is responsible for building up the entire HTTP URL for sending search request to GSA, as per the scope selected by the user, and the serve method(if configured) i.e. public/ public and secure.

# Specific changes

Fix Details
=======================

GSASearchArea.ascx

• Modified the code for SendSearchRequesttoGSA javascript function. The javascript function now internally fires a GET request to the GSA, by making use of ‘window.document.location.href’ statement. 

GSASearchresults – 

• Removed the EnableViewStateMac attribute from the Page directive.