# Feature or Bug Fix
GSBS Patch : 4310682:Effective space utilization of results displayed-font size issue

Description of issue - The utilization of screen space should be more effective for the SearchBox results page. Currently users with 1024*768 resolution only see ~3 results in the screen area and have to scroll down to see more. The possible remedies include:
* Reducing the font size. The current font size is too large and can be reduced. -- can be customized using the CSS classes and XSLT variables used in the stylesheet.
* Remove whitespace at the top of the page/results - please refer http://codereview.appspot.com/4517102/
* make the header/top blue area where the SearchBox is displayed more compact - please refer http://codereview.appspot.com/4517102/


# High level list of changes

Modified the font size for the title of search results displayed.

This patch also contains fix for the Buganizer issue "[Issue 4242469] Error when importing search box stylesheet (SPS_frontend.xslt) into GSA".

# Specific changes

SPS_frontend.xslt
-----------------------
o Modified the value for the "res_title_size" XSLT variable to a lower value, so as to decrease the font size for search results title.
o Added the XSLT namepace in the beginning of the document. This is fix for the issue "[Issue 4242469] Error when importing search box stylesheet (SPS_frontend.xslt) into GSA".