# Feature or Bug Fix Support H2 database for user data store
# High level list of changes
Since Connector Manager 2.8 will support H2 as the default DB, the SP connector needs to support it

As part of the same need to ensure all SQLs are correct and work with H2



# Specific flow/logic that the reviewer needs to focus on sharepoint/source/java/com/google/enterprise/connector/sharepoint/sql.properties

The DDL statement to create the single User data store table fails as the "CHARSET" clause is not supported by H2.

Dropping the same from the DDL statement. The same sql.properties can always be overriden with customized SQLs and shared in WEB-INF/classes directory

#Testing
1. End-to-end integration testing

Review details:
http://codereview.appspot.com/4551087/