# Feature or Bug Fix
ACLs for users/ groups with Limited Access Permission setting on an entity should not be sent to the GSA



# High level list of changes


Modified logic for 'FetchRoleAssignmentsForAcl' function to check the existence of 'limited access' permission setting on an entity for a principal.

# Specific changes

GssAcl.asmx
-----------------------
o Modified code for 'FetchRoleAssignmentsForAcl' function, for handling the case of Limited Access Permissions.
o Logic consists of retrieving the Role Definitions for the current web site having Limited Access permissions. 
o The acls for the user/ group having limited access on an entity will not be send to the GSA. This can be done by avoiding adding such <principal, SharePoint> entries to the aceMap dicitionary object.