# Feature or Bug Fix
[Issue 4274215] Make filter parameter configurable in the search box(2010)

Google search uses two types of automatic filters:
* Duplicate Snippet Filter 
* Duplicate Directory Filter 

These filters can be enabled or disabled by using the filter parameter.

# High level list of changes

Added filter search parameter to search request sent to GSA, so as to apply the filtering on search results.

# Specific changes

GSASearchresults.aspx
-----------------------
o Added filter search parameter to the search request sent to the GSA.
o The value for the filter parameter will be retrieved from the web.config file.
o The value for filter parameter will be written to the web.config file at the time of installation of GSARKS installer and can be modified post-installation.


Note - Please note that only the changes from search box code side are done. Changes required on the C# installer project are yet to be done.