#ISSUE Exclude metadata for documents which match the exclude metadata patterns
#Changes

Modified SPDocument.java for excluding the metadata properties which come under
excluded patterns in get property method .Added Test Methods for the same.

The actual change is that whatever the meta data attributes that the connector
want to exclude from the feed file while generating feed file, you need to
configure the meta data name in the connector instance.xml file and then restart
the connector service. For example just add a meta data attribute name to the
excluded section in the connectorinstance.xml file like below and you need to
uncomment the excluded_metadata property before that.

<!-- EXCLUDED_METADATA:
        Remove any of the metadata entries specified as <value> ... </value> if
you want them to be indexed in GSA. -->

    <property name="excluded_metadata">
        <set>
        <value>.*parentwebtitle$</value>        
      </set>
    </property>