Added a test for code changes for b/5941470 (Sharepoint SOAP response has duplicate attributes). I added an Axis transport and sender so that tests can now load a SOAP response from a file and pass that to Axis for parsing. This allows us to write tests that load SOAP responses from a file that repros a bug and verify code fixes against these captured responses.
Changes included are:

1) I modified the client WSDD file to include a new tranport handler. This is
needed so that we can use a custom Axis transport and sender to read SOAP
responses from a file instead of making an http request. Transport and sender
handlers are registered on this file. The are created from her at runtime.

2) I added a new Axis transport and sender that is used with 1).

3) I modified SharepointTraversalManager to move the SAX client factory
registration to the SharepointClient constructor.

4) Added a new test for b/5941470 using the SOAP response net capture from the
customer.

Code review: http://codereview.appspot.com/5677083/