Split build.properties into two pieces:
* version.properties
* google-enterprise-connector-sharepoint.properties.sample

The latter should be copied to your home directory without the
".sample", and edited as desired. This avoids the issue of modifying
build.properties but having to keep the modifications out of code
reviews and commit. The only such modifications now are to the copy in
your home directory. This follows the standard practice of most of the
other connectors.

Also changed these files to native line endings.

Before updating to this revision, you should revert any local changes
to build.properties, and after updating, you will need to create a
google-enterprise-connector-sharepoint.properties file in your home
directory with the CONNECTOR_MANAGER_DIR property, at least.

Code review: http://codereview.appspot.com/6034051/