Add namespaces to the AD groups and Sharepoint connectors, part 1 of 2.
This part does all of the wiring and almost none of the work. The
global namespace is used in the AD groups connector, but neither is
used in the SharePoint connector at this point. Part 2 will include
the use of the namespaces in the SharePoint connector.

Code review: http://codereview.appspot.com/5978053/