Very last updates to the 2.8.6 release notes. Added bugs 5963073 and 6341529, based on r1803 and comparing it to r1823.
Ondra has argued that "Bug ID b/6341529 should not really be used
since it's applicable only to [one customer], we should be using
b/5923376 instead."

That doesn't seem to be true, given the changes and references in the
bugs and r1803. b/6341529 refers to issues with cyclic groups, and
b/5923376 is a problem with different groups sharing a CN. Revision
r1803 fixes the issue with cyclic groups in the legacy LDAP code, but
cites b/6341529, and b/5923376 makes no mention of cyclic groups.

Code review: https://codereview.appspot.com/6851075