Followup to r2096 (synchronization on checkConnectivity) to update the changes in r2092. Since I committed r2096 separately in order to merge it into 3.0.6, there are no functional changes in this revision.
I added synchronization to the resetHttpClient helper method, to
ensure correctness in lieu of @GuardedBy or just a comment.

Also reworked a warning message to make it clear it was describing
something that the connector was not doing on purpose, and not
something that the connector administrator should avoid doing.

Code review: https://codereview.appspot.com/7706047/