Fixed issue 6553141: Rewrite URL properties are broken since 2.8.0.
The reWriteDisplayUrlUsingAliasMappingRules and
reWriteRecordUrlUsingAliasMappingRules properties were broken in
r1248, which is a merge from /changes/db_impl to /trunk. It looks like
it was caused by a bad merge from the trunk to db_impl in r996. It has
part of the original change for rewriting URLs in r986 but not the
calls to actually pass these parameters from SharepointConnector to
SharepointClientContext (from whence they are passed to
SPDocumentList). So the default values in SharepointClientContext are
always used, and these are true for display URLs and false for record
URLs. The fix is just adding back the following lines to
SharepointConnector.init:

sharepointClientContext.setReWriteDisplayUrlUsingAliasMappingRules(
    reWriteDisplayUrlUsingAliasMappingRules);
sharepointClientContext.setReWriteRecordUrlUsingAliasMappingRules(
    reWriteRecordUrlUsingAliasMappingRules);

I didn't do much in the unit tests, just added what I needed.

Other changes:

* Fixed a typo in a private field name.

Code review: https://codereview.appspot.com/10384044/