Fix issue 8595321: Connector always does full crawl when traversing the AD server.
The problem is that startTraversal always returns null, so the CM
never has a checkpoint value, and it always calls startTraversal. The
fix is to return an empty document list with a non-null checkpoint
from startTraversal.

I added AdGroupsTraversalManagerTest for the test case, rather than
adding to the pile of possibly-related tests in AdGroupsConnectorTest.

Other changes:

* The tests were missing spring.jar and its friends from the classpath.

* AdGroupsConnectorTest.java had the wrong logger name.

Code review: https://codereview.appspot.com/44880043/