Remove the use of SocialCollectionHandler. The user profile collection must now be created manually. The collection name is still a configuration property, since it is used in the docids.
Remove the GSA host, port, admin username, and admin password properties.

The Connector for SharePoint User Profiles 4.0 (aka adaptor)
intentionally requires the collection to be created manually, so this
change is forward-compatible if not backward-compatible.

This is part of removing the GSA Admin API so that we can upgrade
Guava. The GSA Admin API depends on the GData library, which in turn
depends on Guava version 10 or earlier.

Code review: https://codereview.appspot.com/118460044/